+++ libvipster/cell.h
@@ -9,11 +9,11 @@ namespace Vipster {
     enum class CdmFmt { Bohr, Angstrom };
 
     struct CellData {
-        bool                            enabled{false};
-        float                           dimBohr{1};
-        float                           dimAngstrom{bohrrad};
-        Mat                             cellvec{{{{1,0,0}},{{0,1,0}},{{0,0,1}}}};
-        Mat                             invvec{{{{1,0,0}},{{0,1,0}},{{0,0,1}}}};
+        bool    enabled{false};
+        double  dimBohr{1};
+        double  dimAngstrom{bohrrad};
+        Mat     cellvec{{{{1,0,0}},{{0,1,0}},{{0,0,1}}}};
+        Mat     invvec{{{{1,0,0}},{{0,1,0}},{{0,0,1}}}};
     };
 }
